Rename faaCifp* files to provider-neutral identifiers
The runway-map path moved off FAA CIFP last commit, so the surrounding
files no longer parse CIFP-specific data exclusively — they parse
procedure and runway records that happen to come from FAA CIFP today,
but could come from another provider tomorrow. Rename so the names
describe the data they handle, not the current upstream:

  faaCifpLiveDataClient.js     → procedureSourceClient.js
  faaCifpLiveDataModel.js      → procedureSourceModel.js
  faaCifpProcedureModel.js     → procedureRecordModel.js
  faaCifpRunwayModel.js        → runwayRecordModel.js
  faaCifpRunwayProcedureModel.js → runwayProcedureModel.js

Function renames follow the same pattern:
  parseFaaCifpProcedures      → parseProcedureRecords
  parseFaaCifpRunways         → parseRunwayRecords
  extractFaaCifpTextFromZip   → extractProcedureTextFromZip
  discoverActiveCifpRelease   → discoverActiveProcedureRelease

Config:
  FAA_CIFP_CONFIG → PROCEDURE_DATA_CONFIG

Drop the buildAirportRunwayMap alias outright (no backward-compat shim).

GeoJSON `source: "FAA CIFP"` attributions are kept — those are factual
data lineage, not API surface.
